enforce maximum length on workspace_id
This helps applications to not make assumptions that are false. 64 characters is a common limit that is not too restrictive for users/use-cases. In most cases special characters are not allowed at the start of identifiers.

By having these constraints it is easier to map workspace_id's to other resources that might impose similar constraints.